requirements.txt
webargs~=5.5.3
deepaas>=2.1.0
torch==2.4.0
pillow==10.4.0
torchvision==0.19.0
numpy~=2.0.0
ipdb
tqdm
torcheval
pandas
tensorboardX

#torch-tb-profiler

tensorboard
tensorflow
